+++ chat-service/pom.xml
@@ -140,7 +140,7 @@
 		<dependency>
 			<groupId>net.logstash.logback</groupId>
 			<artifactId>logstash-logback-encoder</artifactId>
-			<version>6.6</version>
+			<version>7.2</version>
 		</dependency>
 		<dependency>
 			<groupId>com.epages</groupId>
