--- chat-service/src/main/java/webrtc/v1/config/AsyncConfig.java
@@ -1,21 +1,0 @@
-package webrtc.v1.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@Configuration
-@EnableAsync
-public class AsyncConfig {
-
-    @Bean
-    public Executor myPool() {
-        ThreadPoolTaskExecutor threadPoolTaskExecutor = new ThreadPoolTaskExecutor();
-        threadPoolTaskExecutor.setCorePoolSize(3);    // 기본 스레드 수
-        threadPoolTaskExecutor.setMaxPoolSize(10);     // 최대 스레드 수
-        return threadPoolTaskExecutor;
-    }
-}
