+++ chat-service/src/main/java/webrtc/v1/config/redis/RedisConfig.java
@@ -1,10 +1,15 @@
-package webrtc.v1.config;
+package webrtc.v1.config.redis;
 
+import java.time.Duration;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cache.CacheManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
+import org.springframework.data.redis.cache.RedisCacheConfiguration;
+import org.springframework.data.redis.cache.RedisCacheManager;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
@@ -14,7 +19,9 @@ import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
+import org.springframework.data.redis.serializer.RedisSerializationContext;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 import webrtc.v1.channel.entity.Channel;
 import webrtc.v1.utils.pubsub.RedisSubscriberImpl;
@@ -24,6 +31,8 @@ import webrtc.v1.utils.pubsub.RedisSubscriberImpl;
 @EnableRedisRepositories
 public class RedisConfig {
 
+  @Autowired
+  private RedisConnectionFactory redisConnectionFactory;
   @Value("${spring.redis.port}")
   private int port;
 
@@ -79,5 +88,17 @@ public class RedisConfig {
   @Bean
   public ValueOperations<String, Object> opsValueOperation(RedisTemplate<String, Object> redisTemplate) {
       return redisTemplate.opsForValue();
+  }
+
+
+
+  @Bean
+  public CacheManager redisCacheManager() {
+    RedisCacheConfiguration redisCacheConfiguration = RedisCacheConfiguration.defaultCacheConfig()
+        .serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(new StringRedisSerializer()))
+        .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new GenericJackson2JsonRedisSerializer()))
+        .entryTtl(Duration.ofSeconds(30 * 60));
+
+    return RedisCacheManager.RedisCacheManagerBuilder.fromConnectionFactory(redisConnectionFactory).cacheDefaults(redisCacheConfiguration).build();
   }
 }
