+++ chat-service/src/main/java/webrtc/v1/config/security/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package webrtc.v1.config;
+package webrtc.v1.config.security;
 
 
 import lombok.RequiredArgsConstructor;
@@ -17,6 +17,8 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import webrtc.v1.config.jwt.CustomPasswordEncoder;
+import webrtc.v1.config.jwt.JwtAuthenticationEntryPoint;
 import webrtc.v1.utils.jwt.service.JwtRequestFilter;
 
 import java.util.List;
