+++ chat-service/src/main/java/webrtc/v1/config/socket/WebSockConfig.java
@@ -1,4 +1,4 @@
-package webrtc.v1.config;
+package webrtc.v1.config.socket;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
