+++ chat-service/src/test/resources/logback-test.xml
@@ -1,0 +1,1 @@
+<configuration />
